==> 4_bodylotion.ipynb <==
import requests
from bs4 import BeautifulSoup

Product_Name = []
Discount_Prices = []
Actual_Prices = []
Weight_Flavour = []
Discount_Percentage = []
Rating = []
Users = []

for i in range(2,150):
    url = 'https://www.flipkart.com/search?q=body+lotion&sid=g9b%2Cema%2C5la%2C1ke&as=on&as-show=on&otracker=AS_QueryStore_OrganicAutoSuggest_1_4_na_na_ps&otracker1=AS_QueryStore_OrganicAutoSuggest_1_4_na_na_ps&as-pos=1&as-type=RECENT&suggestionId=body+lotion%7CMoisturizer&requestId=0e2dd518-7ca9-451d-87d9-b7f5d7ed98bc&as-searchtext=body&page='+str(i)
    r=requests.get(url)
    soup = BeautifulSoup(r.text,'lxml')
    box = soup.find('div',class_ = 'DOjaWF gdgoEp')
    
    #Product_name
    names = box.find_all('a', class_ = 'wjcEIp')
    for i in names:
        name = i.text
        Product_Name.append(name)
    #print(Product_name)
    
    #Discount_Prices
    prices = box.find_all('div',class_ = 'Nx9bqj')
    for i in prices:
        name = i.text
        Discount_Prices.append(name)
    #print(Discount_Prices)
    
    #Actual Price
    act_prices = box.find_all('div',class_ = 'yRaY8j')
    for i in act_prices:
        actual_prices = i.text
        Actual_Prices.append(actual_prices)
    #print(Actual_Prices)
    
    #WWeight_Flavour
    flv = box.find_all('div',class_ = 'NqpwHC')
    for i in flv:
        flavour = i.text
        Weight_Flavour.append(flavour)
    #print(Weight_Flavour)
    
    #DDiscount_Percentage
    percent = box.find_all('div',class_='UkUFwK')
    for i in percent:
        percentage = i.text
        Discount_Percentage.append(percentage)
    #print(Discount_Percentage)
    
    #Rating
    rate = box.find_all('div',class_ = 'XQDdHH')
    for i in rate:
        rating = i.text
        Rating.append(rating)
    #print(Rating)
    
    #users
    use = box.find_all('span',class_ = 'Wphh3N')
    for i in use:
        uses = i.text
        Users.append(uses)
    #print(Users)




import pandas as pd
print("Product_Name:", len(Product_Name))
print("Discount_Prices:", len(Discount_Prices))
print("Actual_Prices:", len(Actual_Prices))
print("Weight_Flavour:", len(Weight_Flavour))
print("Discount_Percentage:", len(Discount_Percentage))
print("Rating:", len(Rating))
print("Users:", len(Users))

# Find the minimum length
min_len = min(len(Product_Name), len(Discount_Prices), len(Actual_Prices),
              len(Weight_Flavour), len(Discount_Percentage), len(Rating), len(Users))

# Trim all lists to that length
df = pd.DataFrame({
    'Product Name': Product_Name[:min_len],
    'Discount Prices': Discount_Prices[:min_len],
    'Actual Prices': Actual_Prices[:min_len],
    'Weight/Type': Weight_Flavour[:min_len],
    'Discount Percentage(%)': Discount_Percentage[:min_len],
    'Rating': Rating[:min_len],
    'Users': Users[:min_len]
})
df.to_csv('4_health_supplement.csv',index=False)
df
