==> 5_Data Analysis.ipynb <==
# Analysis the Data 

## Merging & Joining Data

import pandas as pd

df1 = pd.read_csv('1_health_supplement.csv')  #data page 1

df1

df2 = pd.read_csv('2_health_supplement.csv') #data page 2

df2

df3 = pd.read_csv('3_health_supplement.csv') #data page 3

df3

df4 = pd.read_csv('4_health_supplement.csv')  #data page 4

df4

df5 = pd.concat([df1,df2,df3,df4])   #concat all data page (1,2,3,4) in 5 all data in here 

df5   #it is the final data upgrade ,onward what we perform operation by using this data 

df5.to_csv('5_total_health_supplement.csv',index=False)   #push our final data(df5) to a csv file 

## Exploring the Data with Analysis

df = pd.read_csv('total_health_supplement.csv')

df.head()   #it give first 5 rows

df.tail()  #iit gives last 5 rows

df.describe()   #describe the data

df.info()   #all about information data frame

df.columns   #all columns name

df.shape   #(rews,columns)

## Data Selection and Filtering

df

#accesing only ono columns
df['Product Name']

#access one cloumn with periticular index
df['Product Name'][3]  

type(df['Product Name'])
#giving the type of data 

#acessing 2 columns
df[['Product Name','Discount Prices(?)']]

df.loc[1]   #give all about data of index location 1

df.iloc[1]

df.loc[4,'Discount Percentage(%)']   #here we said like give the data  index 7 on discount columns

#by using iloc we access data by using only index number
df.iloc[4,4]   #4 index ,#4 stand for column with index wise

df.loc[0:3,['Discount Percentage(%)','Product Name']]   #said like give data of index from 0 to 3 of column Discount Percentage(%)','Product Name

df.iloc[0:,0:3]   #by acess by using index value by giveind row index columns index

#we want the like give data which rating is grater than 4.5
df[df['Rating']>4.5]

#we want the like give data which rating is grater than 4.5 give that product name only
df[df['Rating']>4.5]['Product Name']

#lets use and functions
#here we said like give highest rating product which has 85% Discount Percentage(%)
df[(df['Rating']>4.5)&(df['Discount Percentage(%)']>'85% off')]    #here we probkem becz of in our funtion we str charecter ,undertsand when you use you also take int data.


df.query("Users > '100000'").head(1)   #here we said like give data which has greater than 100000 user Users columns ,give fist one columns

## Data Cleaning & Preprocessing

import pandas as pd
df = pd.read_csv('total_health_supplement.csv')

#lets check which are column are interger or string
df.dtypes
#shit all value which containis number data that store in number format

# Detect Which Values Are Not Proper Numbers
import pandas as pd

# Assuming df is already loaded
cols_to_check = ['Discount Prices(?)', 'Actual Prices(?)', 'Discount Percentage(%)', 'Users']

for col in cols_to_check:
    print(f"\nChecking column: {col}")
    invalid_rows = df[col].apply(lambda x: pd.to_numeric(str(x).replace(",", "").replace("% off", "").strip(), errors='coerce')).isna()
    if invalid_rows.any():
        print(f"Non-numeric values found in column '{col}':")
        print(df.loc[invalid_rows, col])
    else:
        print("All values are numeric (after cleaning).")


#Convert to Numeric After Cleaning
df['Discount Prices(?)'] = pd.to_numeric(df['Discount Prices(?)'].str.replace(",", "").str.strip(), errors='coerce')
df['Actual Prices(?)'] = pd.to_numeric(df['Actual Prices(?)'].str.replace(",", "").str.strip(), errors='coerce')

# Remove '% off' and convert to numeric
df['Discount Percentage(%)'] = df['Discount Percentage(%)'].str.replace('% off', '').str.strip()
df['Discount Percentage(%)'] = pd.to_numeric(df['Discount Percentage(%)'], errors='coerce')

# Remove commas in Users and convert to integer
df['Users'] = df['Users'].str.replace(",", "").str.strip()
df['Users'] = pd.to_numeric(df['Users'], errors='coerce').astype("Int64")  # Keeps NaN support


df

#lests check again
print(df.dtypes)


#lets check is there is null 
df.isnull()

#method-2 it is best way to see is in this data set there is null or what
df.isnull().sum()
#it gives 0 that means  there is no null value

#lets check is there is any duplicated
df.duplicated()
#it gives false there is no duplicate

## Data Transformation

df

#sorting the value of Discount price and Actual price
df.sort_values(['Discount Prices(?)','Actual Prices(?)'])

df.describe()

df.info()

#we finssh all operation in our data set.now our data set is 100% capable for our Machine Learning
#lets store our clean data set 
df.to_csv('final_health_supplement.csv',index=False)

we do not want to perform all operation peroform in this data set .bcz of we use Machine Learning Forwards

in on next section i starting Visualizing Data for our best Understanding

# Visualzing the Data
